requirements.txt
# Blofin Trading Bot Dependencies

requests>=2.28.0
numpy>=1.24.0
pandas>=1.5.0

# Optional: WebSocket market data stream
websocket-client>=1.8.0

# Optional: For better logging
# colorama>=0.4.6
